+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `LimitOrderBook` for filling limit orders
-- Add `LimitOrderFeeVault` for distribute keeper rewards for limit orders
+- Add `LimitOrderRewardVault` for distribute keeper rewards for limit orders
 
 ## [1.3.5] - 2022-03-18